requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
twilio>=8.10.0
openai>=1.6.0
httpx>=0.25.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
